split tests to work around missing nextTick
